refactor(payment): PaymentCompensationConsumer 보상 흐름 강화 - @Transactional 제거 (HTTP 호출을 DB 트랜잭션으로 감싸지 않음) - Two-phase 패턴 적용: saveIfAbsent → 보상 수행 → markCompleted - cancelOrderAndRestoreStock: !cancelled 시 return 제거, 재고 복구 계속 진행 - refundPoints()를 RefundPointsService로 추출 (Spring AOP proxy self-call 해결)
